% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmaiIntro-package.R
\docType{package}
\name{dmaiIntro}
\alias{dmaiIntro}
\alias{dmaiIntro-package}
\title{Divisia Monetary Aggregates Index}
\description{
The \code{dmai} package provides functionalities to calculate
Divisia monetary aggregates index as given in  Barnett, W. A. (1980).
}
\references{
Barnett, W. A. (1980).
 Economic Monetary Aggregates: An Application of Aggregation and Index Number Theory.
 \emph{Journal of Econometrics}.
 \strong{14}(1):11-48. (\href{https://www.sciencedirect.com/science/article/pii/0304407680900706}{https://www.sciencedirect.com/science/article/pii/0304407680900706}).
}
\author{
\enumerate{
         \item Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item Ahmad Nadeem (\email{Ahmed.Nadeem@sbp.org.pk})
         }
}
